Add files via upload
The Image Filters and Effects project focuses on the application of various image transformations and effects to enhance, manipulate, or stylize digital images. By utilizing techniques from image processing and computer graphics, this project provides tools to modify the appearance of an image through a series of filters and effects.

Key Features and Filters:

Grayscale Conversion:
Converts the image from full color to shades of gray. This is done by removing all the color information and representing the image solely based on its brightness or intensity levels.

Gaussian Blur:
A smoothing filter that blurs the image by averaging the surrounding pixels, effectively reducing detail and noise. This filter is often used to reduce high-frequency noise in images or to create a soft, dreamy effect.